feat(net): remove network address
