+++ crates/net/src/types/mod.rs
@@ -1,6 +1,6 @@
 //! Common types for the Networking Crate.
 
-pub mod address;
 pub mod enr;
 pub mod envelope;
 pub mod payload;
+pub mod peer;
